Split file_test.go into multiple files
