+++ operator/builtin/input/file/file_test.go
@@ -18,9 +18,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
-	"log"
-	"math/rand"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -29,7 +26,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/observiq/nanojack"
 	"github.com/stretchr/testify/require"
 
 	"github.com/open-telemetry/opentelemetry-log-collection/entry"
@@ -37,82 +33,6 @@ import (
 	"github.com/open-telemetry/opentelemetry-log-collection/operator/helper"
 	"github.com/open-telemetry/opentelemetry-log-collection/testutil"
 )
-
-func newDefaultConfig(tempDir string) *InputConfig {
-	cfg := NewInputConfig("testfile")
-	cfg.PollInterval = helper.Duration{Duration: 50 * time.Millisecond}
-	cfg.StartAt = "beginning"
-	cfg.Include = []string{fmt.Sprintf("%s/*", tempDir)}
-	cfg.OutputIDs = []string{"fake"}
-	return cfg
-}
-
-func newTestFileOperator(t *testing.T, cfgMod func(*InputConfig), outMod func(*testutil.FakeOutput)) (*InputOperator, chan *entry.Entry, string) {
-	fakeOutput := testutil.NewFakeOutput(t)
-	if outMod != nil {
-		outMod(fakeOutput)
-	}
-
-	tempDir := testutil.NewTempDir(t)
-
-	cfg := newDefaultConfig(tempDir)
-	if cfgMod != nil {
-		cfgMod(cfg)
-	}
-	ops, err := cfg.Build(testutil.NewBuildContext(t))
-	require.NoError(t, err)
-	op := ops[0]
-
-	err = op.SetOutputs([]operator.Operator{fakeOutput})
-	require.NoError(t, err)
-
-	return op.(*InputOperator), fakeOutput.Received, tempDir
-}
-
-func openFile(t testing.TB, path string) *os.File {
-	file, err := os.OpenFile(path, os.O_CREATE|os.O_RDWR, 0777)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = file.Close() })
-	return file
-}
-
-func openTemp(t testing.TB, tempDir string) *os.File {
-	return openTempWithPattern(t, tempDir, "")
-}
-
-func reopenTemp(t testing.TB, name string) *os.File {
-	return openTempWithPattern(t, filepath.Dir(name), filepath.Base(name))
-}
-
-func openTempWithPattern(t testing.TB, tempDir, pattern string) *os.File {
-	file, err := ioutil.TempFile(tempDir, pattern)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = file.Close() })
-	return file
-}
-
-func getRotatingLogger(t testing.TB, tempDir string, maxLines, maxBackups int, copyTruncate, sequential bool) *log.Logger {
-	file, err := ioutil.TempFile(tempDir, "")
-	require.NoError(t, err)
-	require.NoError(t, file.Close()) // will be managed by rotator
-
-	rotator := nanojack.Logger{
-		Filename:     file.Name(),
-		MaxLines:     maxLines,
-		MaxBackups:   maxBackups,
-		CopyTruncate: copyTruncate,
-		Sequential:   sequential,
-	}
-
-	t.Cleanup(func() { _ = rotator.Close() })
-
-	return log.New(&rotator, "", 0)
-}
-
-func writeString(t testing.TB, file *os.File, s string) {
-	_, err := file.WriteString(s)
-	require.NoError(t, err)
-}
 
 func TestBuild(t *testing.T) {
 	t.Parallel()
@@ -1157,75 +1077,6 @@ func TestFileReader_FingerprintUpdated(t *testing.T) {
 	require.Equal(t, []byte("testlog1\n"), reader.Fingerprint.FirstBytes)
 }
 
-func stringWithLength(length int) string {
-	charset := "abcdefghijklmnopqrstuvwxyz"
-	b := make([]byte, length)
-	for i := range b {
-		b[i] = charset[rand.Intn(len(charset))]
-	}
-	return string(b)
-}
-
-func waitForOne(t *testing.T, c chan *entry.Entry) *entry.Entry {
-	select {
-	case e := <-c:
-		return e
-	case <-time.After(time.Second):
-		require.FailNow(t, "Timed out waiting for message")
-		return nil
-	}
-}
-
-func waitForN(t *testing.T, c chan *entry.Entry, n int) []string {
-	messages := make([]string, 0, n)
-	for i := 0; i < n; i++ {
-		select {
-		case e := <-c:
-			messages = append(messages, e.Body.(string))
-		case <-time.After(time.Second):
-			require.FailNow(t, "Timed out waiting for message")
-			return nil
-		}
-	}
-	return messages
-}
-
-func waitForMessage(t *testing.T, c chan *entry.Entry, expected string) {
-	select {
-	case e := <-c:
-		require.Equal(t, expected, e.Body.(string))
-	case <-time.After(time.Second):
-		require.FailNow(t, "Timed out waiting for message", expected)
-	}
-}
-
-func waitForMessages(t *testing.T, c chan *entry.Entry, expected []string) {
-	receivedMessages := make([]string, 0, len(expected))
-LOOP:
-	for {
-		select {
-		case e := <-c:
-			receivedMessages = append(receivedMessages, e.Body.(string))
-		case <-time.After(time.Second):
-			break LOOP
-		}
-	}
-
-	require.ElementsMatch(t, expected, receivedMessages)
-}
-
-func expectNoMessages(t *testing.T, c chan *entry.Entry) {
-	expectNoMessagesUntil(t, c, 200*time.Millisecond)
-}
-
-func expectNoMessagesUntil(t *testing.T, c chan *entry.Entry, d time.Duration) {
-	select {
-	case e := <-c:
-		require.FailNow(t, "Received unexpected message", "Message: %s", e.Body.(string))
-	case <-time.After(d):
-	}
-}
-
 func TestEncodings(t *testing.T) {
 	t.Parallel()
 	cases := []struct {
@@ -1311,64 +1162,6 @@ func TestEncodings(t *testing.T) {
 	}
 }
 
-type fileInputBenchmark struct {
-	name   string
-	config *InputConfig
-}
-
-func BenchmarkFileInput(b *testing.B) {
-	cases := []fileInputBenchmark{
-		{
-			"Default",
-			NewInputConfig("test_id"),
-		},
-		{
-			"NoFileName",
-			func() *InputConfig {
-				cfg := NewInputConfig("test_id")
-				cfg.IncludeFileName = false
-				return cfg
-			}(),
-		},
-	}
-
-	for _, tc := range cases {
-		b.Run(tc.name, func(b *testing.B) {
-			tempDir := testutil.NewTempDir(b)
-			path := filepath.Join(tempDir, "in.log")
-
-			cfg := tc.config
-			cfg.OutputIDs = []string{"fake"}
-			cfg.Include = []string{path}
-			cfg.StartAt = "beginning"
-
-			ops, err := cfg.Build(testutil.NewBuildContext(b))
-			require.NoError(b, err)
-			op := ops[0]
-
-			fakeOutput := testutil.NewFakeOutput(b)
-			err = op.SetOutputs([]operator.Operator{fakeOutput})
-			require.NoError(b, err)
-
-			err = op.Start(testutil.NewMockPersister("test"))
-			defer op.Stop()
-			require.NoError(b, err)
-
-			file, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE, 0666)
-			require.NoError(b, err)
-
-			for i := 0; i < b.N; i++ {
-				file.WriteString("testlog\n")
-			}
-
-			b.ResetTimer()
-			for i := 0; i < b.N; i++ {
-				<-fakeOutput.Received
-			}
-		})
-	}
-}
-
 // TestExclude tests that a log file will be excluded if it matches the
 // glob specified in the operator.
 func TestExclude(t *testing.T) {
@@ -1410,15 +1203,4 @@ func TestExcludeDuplicates(t *testing.T) {
 
 	matches := getMatches(includes, excludes)
 	require.ElementsMatch(t, matches, paths[2:3])
-}
-
-// writes file with the specified file names and returns their full paths in order
-func writeTempFiles(tempDir string, names []string) []string {
-	result := make([]string, 0, len(names))
-	for _, name := range names {
-		path := filepath.Join(tempDir, name)
-		ioutil.WriteFile(path, []byte(name), 0755)
-		result = append(result, path)
-	}
-	return result
 }
